create DebugInspector to create debug command
