+++ Source/Scene/Scene.js
@@ -34,11 +34,8 @@ import ClearCommand from "../Renderer/ClearCommand.js";
 import ComputeEngine from "../Renderer/ComputeEngine.js";
 import Context from "../Renderer/Context.js";
 import ContextLimits from "../Renderer/ContextLimits.js";
-import DrawCommand from "../Renderer/DrawCommand.js";
 import Pass from "../Renderer/Pass.js";
 import RenderState from "../Renderer/RenderState.js";
-import ShaderProgram from "../Renderer/ShaderProgram.js";
-import ShaderSource from "../Renderer/ShaderSource.js";
 import BrdfLutGenerator from "./BrdfLutGenerator.js";
 import Camera from "./Camera.js";
 import Cesium3DTilePass from "./Cesium3DTilePass.js";
@@ -72,6 +69,7 @@ import SunLight from "./SunLight.js";
 import SunPostProcess from "./SunPostProcess.js";
 import TweenCollection from "./TweenCollection.js";
 import View from "./View.js";
+import DebugInspector from "./DebugInspector.js";
 
 var requestRenderAfterFrame = function (scene) {
   return function () {
@@ -262,6 +260,7 @@ function Scene(options) {
   this._postRender = new Event();
 
   this._minimumDisableDepthTestDistance = 0.0;
+  this._debugInspector = new DebugInspector();
 
   /**
    * Exceptions occurring in <code>render</code> are always caught in order to raise the
@@ -1964,143 +1963,6 @@ Scene.prototype.isVisible = function (command, cullingVolume, occluder) {
   );
 };
 
-function getAttributeLocations(shaderProgram) {
-  var attributeLocations = {};
-  var attributes = shaderProgram.vertexAttributes;
-  for (var a in attributes) {
-    if (attributes.hasOwnProperty(a)) {
-      attributeLocations[a] = attributes[a].index;
-    }
-  }
-
-  return attributeLocations;
-}
-
-function createDebugFragmentShaderProgram(command, scene, shaderProgram) {
-  var context = scene.context;
-  var sp = defaultValue(shaderProgram, command.shaderProgram);
-  var fs = sp.fragmentShaderSource.clone();
-
-  var targets = [];
-  fs.sources = fs.sources.map(function (source) {
-    source = ShaderSource.replaceMain(source, "czm_Debug_main");
-    var re = /gl_FragData\[(\d+)\]/g;
-    var match;
-    while ((match = re.exec(source)) !== null) {
-      if (targets.indexOf(match[1]) === -1) {
-        targets.push(match[1]);
-      }
-    }
-    return source;
-  });
-  var length = targets.length;
-
-  var newMain = "";
-  if (scene.debugShowCommands) {
-    newMain += "uniform vec3 debugShowCommandsColor;\n";
-  }
-  if (scene.debugShowFrustums) {
-    newMain += "uniform vec3 debugShowFrustumsColor;\n";
-  }
-
-  newMain += "void main() \n" + "{ \n" + "    czm_Debug_main(); \n";
-
-  var i;
-  if (scene.debugShowCommands) {
-    if (length > 0) {
-      for (i = 0; i < length; ++i) {
-        newMain +=
-          "    gl_FragData[" +
-          targets[i] +
-          "].rgb *= debugShowCommandsColor;\n";
-      }
-    } else {
-      newMain += "    gl_FragColor.rgb *= debugShowCommandsColor;\n";
-    }
-  }
-
-  if (scene.debugShowFrustums) {
-    if (length > 0) {
-      for (i = 0; i < length; ++i) {
-        newMain +=
-          "    gl_FragData[" +
-          targets[i] +
-          "].rgb *= debugShowFrustumsColor;\n";
-      }
-    } else {
-      newMain += "    gl_FragColor.rgb *= debugShowFrustumsColor;\n";
-    }
-  }
-
-  newMain += "}";
-
-  fs.sources.push(newMain);
-
-  var attributeLocations = getAttributeLocations(sp);
-
-  return ShaderProgram.fromCache({
-    context: context,
-    vertexShaderSource: sp.vertexShaderSource,
-    fragmentShaderSource: fs,
-    attributeLocations: attributeLocations,
-  });
-}
-
-function createDerivedDebugCommand(command, scene, result) {
-  if (!defined(result)) {
-    result = {};
-  }
-
-  var shader;
-  if (defined(result.command)) {
-    shader = result.command.shaderProgram;
-  }
-
-  result.command = DrawCommand.shallowClone(command, result.command);
-
-  // setup debug shader if necessary
-  if (!defined(shader) || result.shaderProgramId !== command.shaderProgram.id) {
-    if (defined(shader)) {
-      shader.destroy();
-    }
-
-    result.command.shaderProgram = createDebugFragmentShaderProgram(
-      command,
-      scene
-    );
-    result.shaderProgramId = command.shaderProgram.id;
-  } else {
-    result.command.shaderProgram = shader;
-  }
-
-  // setup uniform for the debug shader
-  var that = this;
-  var uniformMap = result.command.uniformMap;
-  if (scene.debugShowCommands) {
-    if (!defined(result.debugColor)) {
-      result.debugColor = Color.fromRandom();
-    }
-
-    uniformMap.debugShowCommandsColor = function () {
-      return that.result.debugColor;
-    };
-  }
-
-  if (scene.debugShowFrustums) {
-    // Support up to three frustums.  If a command overlaps all
-    // three, it's code is not changed.
-    var r = command.debugOverlappingFrustums & (1 << 0);
-    var g = command.debugOverlappingFrustums & (1 << 1);
-    var b = command.debugOverlappingFrustums & (1 << 2);
-
-    uniformMap.debugShowFrustumsColor = function () {
-      return new Color(r, g, b);
-    };
-  }
-
-  return result;
-}
-
 var transformFrom2D = new Matrix4(
   0.0,
   0.0,
@@ -2277,12 +2139,12 @@ function executeCommand(command, scene, context, passState, debugFramebuffer) {
   }
 
   if (scene.debugShowCommands || scene.debugShowFrustums) {
-    command.debug = createDerivedDebugCommand(command, scene, command.debug);
-    command.debug.command.execute(scene.context, passState);
+    var debugCommand = scene._debugInspector.createShowFrustumsCommand(
+      scene,
+      command
+    );
+    debugCommand.execute(scene.context, passState);
     return;
-  } else if (defined(command.debug)) {
-    command.debug.command.shaderProgram.destroy();
-    delete command.debug;
   }
 
   if (
